reduced traits
