--- src/YaNco.Abstractions/AffOpt.cs
@@ -1,8 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace Dbosoft.YaNco;
-
-internal static class AffOpt
-{
-    internal const MethodImplOptions mops = MethodImplOptions.AggressiveInlining;
-}
