--- src/YaNco.Abstractions/HasSAPRfcConnection.cs
@@ -1,10 +1,0 @@
-﻿using LanguageExt;
-using LanguageExt.Effects.Traits;
-
-namespace Dbosoft.YaNco;
-
-public interface HasSAPRfcConnection<RT> : HasCancel<RT>
-    where RT : struct, HasCancel<RT>
-{
-    Eff<RT, SAPRfcConnectionIO> RfcConnectionEff { get; }
-}
