+++ src/YaNco.Abstractions/Traits/HasSAPRfcServer.cs
@@ -1,0 +1,10 @@
+﻿using LanguageExt;
+using LanguageExt.Effects.Traits;
+
+namespace Dbosoft.YaNco.Traits;
+
+public interface HasSAPRfcServer<RT>
+    where RT : struct, HasSAPRfc<RT>
+{
+    Eff<RT, SAPRfcServerIO> RfcServerEff { get; }
+}
