+++ src/YaNco.Abstractions/Traits/SAPRfcDataIO.cs
@@ -1,0 +1,6 @@
+﻿namespace Dbosoft.YaNco.Traits;
+
+// ReSharper disable once InconsistentNaming
+public interface SAPRfcDataIO : SAPRfcTypeIO, SAPRfcStructureIO, SAPRfcTableIO, SAPRfcFieldIO
+{
+}
