+++ src/YaNco.Core/ConnectionBuilder.cs
@@ -2,7 +2,9 @@
 using System.Collections.Generic;
 using System.Threading;
 using Dbosoft.YaNco.Live;
+using Dbosoft.YaNco.Traits;
 using LanguageExt;
+using LanguageExt.Effects.Traits;
 
 namespace Dbosoft.YaNco;
 
@@ -70,7 +72,7 @@ public class ConnectionBuilder : ConnectionBuilderBase<ConnectionBuilder, SAPRfc
 /// </summary>
 /// <typeparam name="RT">The runtime to be used for the connection</typeparam>
 public class ConnectionBuilder<RT> : ConnectionBuilderBase<ConnectionBuilder<RT>, RT>
-    where RT : struct, HasSAPRfcFunctions<RT>, HasSAPRfcServer<RT>, HasSAPRfcConnection<RT>, HasSAPRfcLogger<RT>, HasSAPRfcData<RT>, IHasEnvRuntimeSettings
+    where RT : struct, HasSAPRfcServer<RT>, HasSAPRfc<RT>, HasCancel<RT>
 
 {
     public ConnectionBuilder(IDictionary<string, string> connectionParam) : base(connectionParam)
