+++ src/YaNco.Core/Live/LiveSAPRfcConnectionIO.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using Dbosoft.YaNco.Internal;
+using Dbosoft.YaNco.Traits;
 using LanguageExt;
 
 namespace Dbosoft.YaNco.Live;
