+++ src/YaNco.Core/Live/LiveSAPRfcDataIO.cs
@@ -1,4 +1,5 @@
 ﻿using Dbosoft.YaNco.Internal;
+using Dbosoft.YaNco.Traits;
 using Dbosoft.YaNco.TypeMapping;
 using LanguageExt;
 using System;
