+++ src/YaNco.Core/Live/LiveSAPRfcFunctionIO.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Dbosoft.YaNco.Internal;
+using Dbosoft.YaNco.Traits;
 using LanguageExt;
 
 namespace Dbosoft.YaNco.Live;
