+++ src/YaNco.Core/Table.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Dbosoft.YaNco.Traits;
 using LanguageExt;
 
 namespace Dbosoft.YaNco;
